# PLAT-1255 Optimized DomAutoCompleteDisplayStringDeduplicator