measurements with about 11000 business partners
```
before:       75ms 84ms 72ms 69ms 92ms
with this PR: 35ms 30ms 41ms 45ms 37ms
```